crud_flask/requirements.txt
alembic==1.15.2
black==25.1.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
python-dotenv==1.0.0
flake8>=6.0.0